pyproject.toml
[project]
name = "contenidos-inacap"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "rich (>=14.3.3,<15.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "openai (>=2.26.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pypdf (>=6.8.0,<7.0.0)",
    "python-docx (>=1.2.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "contenidos_inacap", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "ruff (>=0.15.5,<0.16.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "black (>=26.3.0,<27.0.0)"
]
